refactor(api): migrate tutorialEvaluation from lib to src
