+++ api/lib/domain/models/index.js
@@ -15,6 +15,7 @@ import { Training } from '../../../src/devcomp/domain/models/Training.js';
 import { TrainingTrigger } from '../../../src/devcomp/domain/models/TrainingTrigger.js';
 import { TrainingTriggerTube } from '../../../src/devcomp/domain/models/TrainingTriggerTube.js';
 import { Tutorial } from '../../../src/devcomp/domain/models/Tutorial.js';
+import { TutorialEvaluation } from '../../../src/devcomp/domain/models/TutorialEvaluation.js';
 import { UserSavedTutorial } from '../../../src/devcomp/domain/models/UserSavedTutorial.js';
 import { Answer } from '../../../src/evaluation/domain/models/Answer.js';
 import { CompetenceEvaluation } from '../../../src/evaluation/domain/models/CompetenceEvaluation.js';
@@ -126,7 +127,6 @@ import { TargetProfileForCreation } from './TargetProfileForCreation.js';
 import { TargetProfileSummaryForAdmin } from './TargetProfileSummaryForAdmin.js';
 import { Thematic } from './Thematic.js';
 import { Tube } from './Tube.js';
-import { TutorialEvaluation } from './TutorialEvaluation.js';
 import { UserCompetence } from './UserCompetence.js';
 import { UserDetailsForAdmin } from './UserDetailsForAdmin.js';
 import { UserOrgaSettings } from './UserOrgaSettings.js';
