+++ api/lib/routes.js
@@ -36,7 +36,6 @@ import * as supOrganizationLearners from './application/sup-organization-learner
 import * as tags from './application/tags/index.js';
 import * as targetProfiles from './application/target-profiles/index.js';
 import * as targetProfilesManagement from './application/target-profiles-management/index.js';
-import * as tutorialEvaluations from './application/tutorial-evaluations/index.js';
 import * as userOrgaSettings from './application/user-orga-settings/index.js';
 import * as users from './application/users/index.js';
 
@@ -79,7 +78,6 @@ const routes = [
   targetProfiles,
   targetProfilesManagement,
   frameworks,
-  tutorialEvaluations,
   userOrgaSettings,
   users,
 ];
