+++ api/src/devcomp/infrastructure/repositories/tutorial-repository.js
@@ -1,7 +1,6 @@
 import _ from 'lodash';
 
 import * as knowledgeElementRepository from '../../../../lib/infrastructure/repositories/knowledge-element-repository.js';
-import * as tutorialEvaluationRepository from '../../../../lib/infrastructure/repositories/tutorial-evaluation-repository.js';
 import * as paginateModule from '../../../../lib/infrastructure/utils/paginate.js';
 import { LOCALE } from '../../../shared/domain/constants.js';
 import { NotFoundError } from '../../../shared/domain/errors.js';
@@ -9,6 +8,7 @@ import { tutorialDatasource } from '../../../shared/infrastructure/datasources/l
 import * as skillRepository from '../../../shared/infrastructure/repositories/skill-repository.js';
 import { Tutorial } from '../../domain/models/Tutorial.js';
 import { TutorialForUser } from '../../domain/read-models/TutorialForUser.js';
+import * as tutorialEvaluationRepository from './tutorial-evaluation-repository.js';
 import * as userSavedTutorialRepository from './user-saved-tutorial-repository.js';
 
 const { FRENCH_FRANCE } = LOCALE;
