+++ api/src/devcomp/infrastructure/serializers/jsonapi/tutorial-attributes.js
@@ -1,5 +1,6 @@
 import { tutorialEvaluationAttributes } from './tutorial-evaluation-attributes.js';
 import { userSavedTutorialAttributes } from './user-saved-tutorial-attributes.js';
+
 const tutorialAttributes = {
   ref: 'id',
   includes: true,
