+++ api/src/devcomp/infrastructure/serializers/jsonapi/tutorial-serializer.js
@@ -1,7 +1,7 @@
 import jsonapiSerializer from 'jsonapi-serializer';
 
-import { tutorialEvaluationAttributes } from '../../../../../lib/infrastructure/serializers/jsonapi/tutorial-evaluation-attributes.js';
-import { userSavedTutorialAttributes } from '../../../../../lib/infrastructure/serializers/jsonapi/user-saved-tutorial-attributes.js';
+import { tutorialEvaluationAttributes } from './tutorial-evaluation-attributes.js';
+import { userSavedTutorialAttributes } from './user-saved-tutorial-attributes.js';
 
 const { Serializer } = jsonapiSerializer;
