+++ api/src/devcomp/routes.js
@@ -1,8 +1,9 @@
 import * as modulesRoutes from './application/modules/index.js';
 import * as passages from './application/passages/index.js';
 import * as trainings from './application/trainings/index.js';
+import * as evaluationTutorials from './application/tutorial-evaluations/index.js';
 import * as userTutorials from './application/user-tutorials/index.js';
 
-const devcompRoutes = [modulesRoutes, passages, trainings, userTutorials];
+const devcompRoutes = [modulesRoutes, passages, trainings, userTutorials, evaluationTutorials];
 
 export { devcompRoutes };
