+++ api/src/shared/infrastructure/datasources/learning-content/index.js
@@ -1,3 +1,4 @@
+import { tutorialDatasource } from '../../../../devcomp/infrastructure/datasources/learning-content/tutorial-datasource.js';
 import { areaDatasource } from './area-datasource.js';
 import { challengeDatasource } from './challenge-datasource.js';
 import { competenceDatasource } from './competence-datasource.js';
@@ -6,7 +7,6 @@ import { frameworkDatasource } from './framework-datasource.js';
 import { skillDatasource } from './skill-datasource.js';
 import { thematicDatasource } from './thematic-datasource.js';
 import { tubeDatasource } from './tube-datasource.js';
-import { tutorialDatasource } from './tutorial-datasource.js';
 
 export {
   areaDatasource,
