+++ api/tests/devcomp/integration/infrastructure/repositories/tutorial-evaluation-repository_test.js
@@ -1,6 +1,6 @@
-import { TutorialEvaluation } from '../../../../lib/domain/models/TutorialEvaluation.js';
-import * as tutorialEvaluationRepository from '../../../../lib/infrastructure/repositories/tutorial-evaluation-repository.js';
-import { databaseBuilder, expect, knex } from '../../../test-helper.js';
+import { TutorialEvaluation } from '../../../../../src/devcomp/domain/models/TutorialEvaluation.js';
+import * as tutorialEvaluationRepository from '../../../../../src/devcomp/infrastructure/repositories/tutorial-evaluation-repository.js';
+import { databaseBuilder, expect, knex } from '../../../../test-helper.js';
 
 describe('Integration | Infrastructure | Repository | tutorialEvaluationRepository', function () {
   let userId, tutorialId, status;
