+++ api/tests/devcomp/unit/application/tutorial-evaluations/index_test.js
@@ -1,7 +1,7 @@
-import * as moduleUnderTest from '../../../../lib/application/tutorial-evaluations/index.js';
-import { tutorialEvaluationsController } from '../../../../lib/application/tutorial-evaluations/tutorial-evaluations-controller.js';
-import { TutorialEvaluation } from '../../../../lib/domain/models/TutorialEvaluation.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
+import * as moduleUnderTest from '../../../../../src/devcomp/application/tutorial-evaluations/index.js';
+import { tutorialEvaluationsController } from '../../../../../src/devcomp/application/tutorial-evaluations/tutorial-evaluations-controller.js';
+import { TutorialEvaluation } from '../../../../../src/devcomp/domain/models/TutorialEvaluation.js';
+import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Router | tutorial-evaluations-router', function () {
   describe('PUT /api/users/tutorials/{tutorialId}/evaluate', function () {
