+++ api/tests/devcomp/unit/application/tutorial-evaluations/tutorial-evaluations-controller_test.js
@@ -1,7 +1,7 @@
-import { tutorialEvaluationsController } from '../../../../lib/application/tutorial-evaluations/tutorial-evaluations-controller.js';
-import { TutorialEvaluation } from '../../../../lib/domain/models/TutorialEvaluation.js';
-import { usecases } from '../../../../lib/domain/usecases/index.js';
-import { expect, hFake, sinon } from '../../../test-helper.js';
+import { tutorialEvaluationsController } from '../../../../../src/devcomp/application/tutorial-evaluations/tutorial-evaluations-controller.js';
+import { TutorialEvaluation } from '../../../../../src/devcomp/domain/models/TutorialEvaluation.js';
+import { usecases } from '../../../../../src/devcomp/domain/usecases/index.js';
+import { expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Controller | Tutorial-evaluations', function () {
   describe('#evaluate', function () {
