+++ api/tests/devcomp/unit/domain/usecases/add-tutorial-evaluation_test.js
@@ -1,7 +1,7 @@
-import { NotFoundError } from '../../../../lib/domain/errors.js';
-import { TutorialEvaluation } from '../../../../lib/domain/models/TutorialEvaluation.js';
-import { addTutorialEvaluation } from '../../../../lib/domain/usecases/add-tutorial-evaluation.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { NotFoundError } from '../../../../../lib/domain/errors.js';
+import { TutorialEvaluation } from '../../../../../src/devcomp/domain/models/TutorialEvaluation.js';
+import { addTutorialEvaluation } from '../../../../../src/devcomp/domain/usecases/add-tutorial-evaluation.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | UseCase | add-tutorial-evaluation', function () {
   let tutorialRepository;
