+++ api/tests/devcomp/unit/infrastructure/datasources/learning-content/tutorial-datasource_test.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 
 import { learningContentCache } from '../../../../../../lib/infrastructure/caches/learning-content-cache.js';
 import { lcms } from '../../../../../../lib/infrastructure/lcms.js';
-import { tutorialDatasource } from '../../../../../../src/shared/infrastructure/datasources/learning-content/tutorial-datasource.js';
+import { tutorialDatasource } from '../../../../../../src/devcomp/infrastructure/datasources/learning-content/tutorial-datasource.js';
 import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Datasource | Learning Content | TutorialDatasource', function () {
