+++ api/tests/unit/infrastructure/serializers/jsonapi/tutorial-evaluation-serializer_test.js
@@ -1,5 +1,5 @@
-import { TutorialEvaluation } from '../../../../../lib/domain/models/TutorialEvaluation.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/tutorial-evaluation-serializer.js';
+import { TutorialEvaluation } from '../../../../../src/devcomp/domain/models/TutorialEvaluation.js';
+import * as serializer from '../../../../../src/devcomp/infrastructure/serializers/jsonapi/tutorial-evaluation-serializer.js';
 import { domainBuilder, expect } from '../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | tutorial-evaluation-serializer', function () {
